Refactor DSL email handler
